add feature#59
