+++ .gitignore
@@ -3,4 +3,4 @@
 !.vscode/tasks.json
 !.vscode/launch.json
 !.vscode/extensions.json
-__pycache__/*
+package/__pycache__/*
