+++ .travis.yml
@@ -10,5 +10,5 @@ env:
 install:
     - pip install -r requirements.txt
 script:
-    - flake8  --ignore E501  main.py
-    - python main.py 1 1 -G localhost root
+    - python main.py -S test
+    - python main.py -G 1 2 localhost root
